refactor: reorganize register state management and improve UI structure in tracks view
