+++ lib/features/Auth/controller/register_cubit/register_cubit.dart
@@ -1,5 +1,5 @@
-import 'package:final_project_iti/features/Auth/controller/cubit/auth_state.dart';
-import 'package:flutter/widgets.dart';
+import 'package:final_project_iti/features/Auth/controller/register_cubit/register_state.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 class RegisterCubit extends Cubit<RegisterState> {
